+++ src/gov/usgs/earthquake/nshmp/data/DoubleData.java
@@ -1,10 +1,8 @@
-package org.opensha.util.data;
+package gov.usgs.earthquake.nshmp.data;
 
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
-import static org.opensha.util.Text.NEWLINE;
-
-import org.opensha.util.Maths;
+import static gov.usgs.earthquake.nshmp.Text.NEWLINE;
 
 import com.google.common.base.Function;
 import com.google.common.base.Strings;
@@ -12,6 +10,8 @@ import com.google.common.collect.Lists;
 import com.google.common.collect.Range;
 import com.google.common.math.DoubleMath;
 import com.google.common.primitives.Doubles;
+
+import gov.usgs.earthquake.nshmp.Maths;
 
 import java.math.BigDecimal;
 import java.util.Arrays;
